// RookieEcommerce.AdminSite/src/api/models/ProductVariantDetailsDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProductVariantDetailsDto
 */
export interface ProductVariantDetailsDto {
    /**
     * 
     * @type {string}
     * @memberof ProductVariantDetailsDto
     */
    id?: string;
    /**
     * 
     * @type {Date}
     * @memberof ProductVariantDetailsDto
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ProductVariantDetailsDto
     */
    modifiedDate?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ProductVariantDetailsDto
     */
    productId?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductVariantDetailsDto
     */
    variantType?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ProductVariantDetailsDto
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ProductVariantDetailsDto
     */
    sku?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ProductVariantDetailsDto
     */
    stockQuantity?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductVariantDetailsDto
     */
    price?: number;
}

/**
 * Check if a given object implements the ProductVariantDetailsDto interface.
 */
export function instanceOfProductVariantDetailsDto(value: object): value is ProductVariantDetailsDto {
    return true;
}

export function ProductVariantDetailsDtoFromJSON(json: any): ProductVariantDetailsDto {
    return ProductVariantDetailsDtoFromJSONTyped(json, false);
}

export function ProductVariantDetailsDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductVariantDetailsDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'createdDate': json['createdDate'] == null ? undefined : (new Date(json['createdDate'])),
        'modifiedDate': json['modifiedDate'] == null ? undefined : (new Date(json['modifiedDate'])),
        'productId': json['productId'] == null ? undefined : json['productId'],
        'variantType': json['variantType'] == null ? undefined : json['variantType'],
        'name': json['name'] == null ? undefined : json['name'],
        'sku': json['sku'] == null ? undefined : json['sku'],
        'stockQuantity': json['stockQuantity'] == null ? undefined : json['stockQuantity'],
        'price': json['price'] == null ? undefined : json['price'],
    };
}

export function ProductVariantDetailsDtoToJSON(json: any): ProductVariantDetailsDto {
    return ProductVariantDetailsDtoToJSONTyped(json, false);
}

export function ProductVariantDetailsDtoToJSONTyped(value?: ProductVariantDetailsDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'createdDate': value['createdDate'] == null ? undefined : ((value['createdDate']).toISOString()),
        'modifiedDate': value['modifiedDate'] == null ? undefined : ((value['modifiedDate'] as any).toISOString()),
        'productId': value['productId'],
        'variantType': value['variantType'],
        'name': value['name'],
        'sku': value['sku'],
        'stockQuantity': value['stockQuantity'],
        'price': value['price'],
    };
}