// RookieEcommerce.AdminSite/src/api/models/ProductVariantDetailsDtoPaginationList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProductVariantDetailsDto } from './ProductVariantDetailsDto';
import {
    ProductVariantDetailsDtoFromJSON,
    ProductVariantDetailsDtoFromJSONTyped,
    ProductVariantDetailsDtoToJSON,
    ProductVariantDetailsDtoToJSONTyped,
} from './ProductVariantDetailsDto';

/**
 * 
 * @export
 * @interface ProductVariantDetailsDtoPaginationList
 */
export interface ProductVariantDetailsDtoPaginationList {
    /**
     * 
     * @type {Array<ProductVariantDetailsDto>}
     * @memberof ProductVariantDetailsDtoPaginationList
     */
    items?: Array<ProductVariantDetailsDto> | null;
    /**
     * 
     * @type {number}
     * @memberof ProductVariantDetailsDtoPaginationList
     */
    pageNumber?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductVariantDetailsDtoPaginationList
     */
    pageSize?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductVariantDetailsDtoPaginationList
     */
    totalCount?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductVariantDetailsDtoPaginationList
     */
    readonly totalPages?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ProductVariantDetailsDtoPaginationList
     */
    readonly hasPreviousPage?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ProductVariantDetailsDtoPaginationList
     */
    readonly hasNextPage?: boolean;
}

/**
 * Check if a given object implements the ProductVariantDetailsDtoPaginationList interface.
 */
export function instanceOfProductVariantDetailsDtoPaginationList(value: object): value is ProductVariantDetailsDtoPaginationList {
    return true;
}

export function ProductVariantDetailsDtoPaginationListFromJSON(json: any): ProductVariantDetailsDtoPaginationList {
    return ProductVariantDetailsDtoPaginationListFromJSONTyped(json, false);
}

export function ProductVariantDetailsDtoPaginationListFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductVariantDetailsDtoPaginationList {
    if (json == null) {
        return json;
    }
    return {
        
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(ProductVariantDetailsDtoFromJSON)),
        'pageNumber': json['pageNumber'] == null ? undefined : json['pageNumber'],
        'pageSize': json['pageSize'] == null ? undefined : json['pageSize'],
        'totalCount': json['totalCount'] == null ? undefined : json['totalCount'],
        'totalPages': json['totalPages'] == null ? undefined : json['totalPages'],
        'hasPreviousPage': json['hasPreviousPage'] == null ? undefined : json['hasPreviousPage'],
        'hasNextPage': json['hasNextPage'] == null ? undefined : json['hasNextPage'],
    };
}

export function ProductVariantDetailsDtoPaginationListToJSON(json: any): ProductVariantDetailsDtoPaginationList {
    return ProductVariantDetailsDtoPaginationListToJSONTyped(json, false);
}

export function ProductVariantDetailsDtoPaginationListToJSONTyped(value?: Omit<ProductVariantDetailsDtoPaginationList, 'totalPages'|'hasPreviousPage'|'hasNextPage'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(ProductVariantDetailsDtoToJSON)),
        'pageNumber': value['pageNumber'],
        'pageSize': value['pageSize'],
        'totalCount': value['totalCount'],
    };
}