// RookieEcommerce.AdminSite/src/api/models/CreateVariantCommand.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateVariantCommand
 */
export interface CreateVariantCommand {
    /**
     * 
     * @type {string}
     * @memberof CreateVariantCommand
     */
    productId?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateVariantCommand
     */
    variantType?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateVariantCommand
     */
    name?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CreateVariantCommand
     */
    stockQuantity?: number;
    /**
     * 
     * @type {number}
     * @memberof CreateVariantCommand
     */
    price?: number;
}

/**
 * Check if a given object implements the CreateVariantCommand interface.
 */
export function instanceOfCreateVariantCommand(value: object): value is CreateVariantCommand {
    return true;
}

export function CreateVariantCommandFromJSON(json: any): CreateVariantCommand {
    return CreateVariantCommandFromJSONTyped(json, false);
}

export function CreateVariantCommandFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateVariantCommand {
    if (json == null) {
        return json;
    }
    return {
        
        'productId': json['productId'] == null ? undefined : json['productId'],
        'variantType': json['variantType'] == null ? undefined : json['variantType'],
        'name': json['name'] == null ? undefined : json['name'],
        'stockQuantity': json['stockQuantity'] == null ? undefined : json['stockQuantity'],
        'price': json['price'] == null ? undefined : json['price'],
    };
}

export function CreateVariantCommandToJSON(json: any): CreateVariantCommand {
    return CreateVariantCommandToJSONTyped(json, false);
}

export function CreateVariantCommandToJSONTyped(value?: CreateVariantCommand | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'productId': value['productId'],
        'variantType': value['variantType'],
        'name': value['name'],
        'stockQuantity': value['stockQuantity'],
        'price': value['price'],
    };
}