// RookieEcommerce.AdminSite/src/api/models/UpdateVariantCommand.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateVariantCommand
 */
export interface UpdateVariantCommand {
    /**
     * 
     * @type {string}
     * @memberof UpdateVariantCommand
     */
    name?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVariantCommand
     */
    stockQuantity?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVariantCommand
     */
    price?: number | null;
}

/**
 * Check if a given object implements the UpdateVariantCommand interface.
 */
export function instanceOfUpdateVariantCommand(value: object): value is UpdateVariantCommand {
    return true;
}

export function UpdateVariantCommandFromJSON(json: any): UpdateVariantCommand {
    return UpdateVariantCommandFromJSONTyped(json, false);
}

export function UpdateVariantCommandFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateVariantCommand {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'stockQuantity': json['stockQuantity'] == null ? undefined : json['stockQuantity'],
        'price': json['price'] == null ? undefined : json['price'],
    };
}

export function UpdateVariantCommandToJSON(json: any): UpdateVariantCommand {
    return UpdateVariantCommandToJSONTyped(json, false);
}

export function UpdateVariantCommandToJSONTyped(value?: UpdateVariantCommand | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'stockQuantity': value['stockQuantity'],
        'price': value['price'],
    };
}